TelegramBotShopController add objectMapper and log.info add v_002_createItems.sql

+++ src/main/java/ru/exampl/bot2/domain/command/StartCommand.java
@@ -10,5 +10,5 @@ public class StartCommand {
     public String chatId;
     public String userName;
 
-    public Integer userId;
+    public Long userId;
 }
